bumped to 1.21

+++ src/main/kotlin/dev/lyzev/schizoid/feature/features/module/modules/render/ModuleToggleableViewClip.kt
@@ -32,7 +32,7 @@ object ModuleToggleableViewClip :
 
     init {
         on<EventClipToSpace> { event ->
-            event.desiredCameraDistance = desiredCameraDistance.toDouble()
+            event.desiredCameraDistance = desiredCameraDistance
         }
     }
 }
